pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "helix-tentacle"
version = "0.1.0"
description = "Helix-Tentacle: External perception microservice for Helix ecosystem"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Jason Milk", email = "jason@helix.eco" }
]
keywords = ["ai", "agent", "perception", "web-scraping", "helix"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "typer>=0.12",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "cachetools>=5.3",
    "tenacity>=9.0",
    "ddgs>=9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.11",
]

[project.scripts]
tentacle = "cli.main:app"

[tool.setuptools]
packages = { find = { where = ["."], include = ["tentacle", "tentacle.*", "cli", "cli.*"] } }

[tool.setuptools.package-data]
tentacle = ["py.typed"]
cli = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"